fix tests again
